Switch back to threads, with improved behavior.
The cooperative multitasking approach also helps in multi-threaded mode,
but for a different reason: you can't fully cancel a thread until the
function returns. Returning every so often (though less often than in
single-threaded mode) makes the experience smoother.

Besides that, this increases the number of threads allowed at once,
which helps when the canvas is split up and multiple sections are being
refilled at once.

This commit concludes the "threads" branch.